refactor: change api to server
